alpha/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alabaster==1.0.0
albucore==0.0.24
albumentations==2.0.6
annotated-types==0.7.0
anyio==4.9.0
asttokens==2.4.1
attrs==24.2.0
audioread==3.0.1
auraloss==0.4.0
babel==2.16.0
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7.1
contourpy==1.3.0
cryptography==44.0.3
cycler==0.12.1
dash==3.0.4
datasets==2.20.0
debugpy==1.8.7
decorator==5.1.1
DeepExperimentManager==0.1.4
dill==0.3.8
docutils==0.21.2
dotenv==0.9.9
einops==0.8.0
et_xmlfile==2.0.0
executing==2.1.0
fastapi==0.115.11
fastjsonschema==2.21.1
filelock==3.16.1
Flask==3.0.3
Flask-SocketIO==5.5.1
flwr==1.20.0
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.5.0
gensim==4.3.3
graphviz==0.20.3
grpcio==1.67.1
grpcio-health-checking==1.62.3
h11==0.14.0
huggingface-hub==0.26.2
idna==3.10
imagesize==1.4.1
imbalanced-learn==0.13.0
imblearn==0.0
importlib_metadata==8.7.0
intel-openmp==2021.4.0
ipykernel==6.29.5
ipython==8.29.0
ipywidgets==8.1.7
iterators==0.0.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.15
keyring==25.5.0
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.2.post1
lightgbm==4.6.0
lightning-utilities==0.11.8
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl==2021.4.0
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
narwhals==1.42.1
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nh3==0.2.20
node2vec==0.5.0
numba==0.60.0
numpy==1.26.4
open3d==0.19.0
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
pillow==11.0.0
pkginfo==1.12.0
platformdirs==4.3.6
plotly==6.1.2
pooch==1.8.2
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==4.25.5
psutil==6.1.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==18.0.0
pyarrow-hotfix==0.6
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.18.0
pyparsing==3.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.1
python-multipart==0.0.20
python-socketio==5.13.0
pytorch-lightning==2.3.0
pytz==2024.2
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==26.2.0
ray==2.31.0
readme_renderer==44.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3986==2.0.0
rich==13.9.4
rpds-py==0.25.1
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
setuptools==75.1.0
shellingham==1.5.4
simple-websocket==1.1.0
simsimd==6.2.1
six==1.16.0
sklearn-compat==0.1.3
smart-open==7.0.5
sniffio==1.3.1
snowballstemmer==2.2.0
soundfile==0.12.1
soxr==0.5.0.post1
Sphinx==8.1.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
starlette==0.46.1
stringzilla==3.12.5
sympy==1.13.3
tbb==2021.13.1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tokenizers==0.19.1
tomli==2.2.1
tomli_w==1.2.0
# torch==2.3.1+cu118
# torchaudio==2.3.1+cu118
# torchinfo==1.8.0
# torchmetrics==1.5.1
# torchvision==0.18.1+cu118
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
transformers==4.42.0
twine==6.0.1
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.146
ultralytics-thop==2.0.14
urllib3==2.2.3
uvicorn==0.34.0
wcwidth==0.2.13
Werkzeug==3.0.6
wheel==0.44.0
widgetsnbextension==4.0.14
wrapt==1.16.0
wsproto==1.2.0
xgboost==2.1.3
xxhash==3.5.0
yarl==1.17.1
zipp==3.23.0